Changed decisecs to uint16_t to prevent overflow with large wait times.

+++ src/mrbw-th.c
@@ -496,7 +496,7 @@ uint16_t system_sleep(uint16_t sleep_decisecs)
 // and the call to this function in the main function
 
 volatile uint8_t ticks;
-volatile uint8_t decisecs;
+volatile uint16_t decisecs;
 
 void initialize100HzTimer(void)
 {
